@@ -4,16 +4,17 @@ version: 2.3.0
 homepage: https://github.com/DeFiCh-WalletApp/DeFiChain-dart
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   bech32: ^0.2.0
-  bip32_defichain: ^2.1.0
-  pointycastle: ^3.0.1
+  bip32_defichain: ^2.2.0
+  pointycastle: ^3.1.1
   hex: ^0.2.0
-  bs58check_dart: ^2.0.0+1
-  meta: ^1.2.4
-  crypto: ^3.0.0
+  bs58check: ^1.0.2
+  meta: ^1.4.0
+  crypto: ^3.0.1
+  collection: ^1.15.0
 
 dev_dependencies:
   test: ^1.0.0
