@@ -1,15 +1,15 @@
 import 'dart:typed_data';
 
 class PaymentData {
-  String name;
-  String address;
-  Uint8List hash;
-  Uint8List output;
-  Uint8List signature;
-  Uint8List pubkey;
-  Uint8List input;
-  List<Uint8List> witness;
-  PaymentData redeem;
+  String? name;
+  String? address;
+  Uint8List? hash;
+  Uint8List? output;
+  Uint8List? signature;
+  Uint8List? pubkey;
+  Uint8List? input;
+  List<Uint8List?>? witness;
+  PaymentData? redeem;
 
   PaymentData(
       {this.name,
