@@ -5,9 +5,9 @@ import '../payments/index.dart' show PaymentData;
 import '../utils/constants/op.dart';
 
 class P2PK {
-  PaymentData data;
-  NetworkType network;
-  P2PK({@required data, network}) {
+  late PaymentData data;
+  NetworkType? network;
+  P2PK({required data, network}) {
     this.network = network ?? bitcoin;
     this.data = data;
     _init();
@@ -15,10 +15,10 @@ class P2PK {
 
   void _init() {
     if (data.output != null) {
-      if (data.output[data.output.length - 1] != OPS['OP_CHECKSIG']) {
+      if (data.output![data.output!.length - 1] != OPS['OP_CHECKSIG']) {
         throw ArgumentError('Output is invalid');
       }
-      if (!isPoint(data.output.sublist(1, -1))) {
+      if (!isPoint(data.output!.sublist(1, -1))) {
         throw ArgumentError('Output pubkey is invalid');
       }
     }
