@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 import 'package:meta/meta.dart';
 import 'package:bip32_defichain/src/utils/ecurve.dart' show isPoint;
-import 'package:bs58check_dart/bs58check.dart' as bs58check;
+import 'package:bs58check/bs58check.dart' as bs58check;
 
 import '../crypto.dart';
 import '../models/networks.dart';
@@ -10,64 +10,72 @@ import '../utils/script.dart' as bscript;
 import '../utils/constants/op.dart';
 
 class P2PKH {
-  PaymentData data;
-  NetworkType network;
-  P2PKH({@required data, network}) {
+  PaymentData? data;
+  late NetworkType network;
+  P2PKH({required data, network}) {
     this.network = network ?? bitcoin;
     this.data = data;
     _init();
   }
   void _init() {
-    data.name = 'p2pkh';
-    if (data.address != null) {
-      _getDataFromAddress(data.address);
+    data!.name = 'p2pkh';
+    if (data!.address != null) {
+      _getDataFromAddress(data!.address!);
       _getDataFromHash();
-    } else if (data.hash != null) {
+    } else if (data!.hash != null) {
       _getDataFromHash();
-    } else if (data.output != null) {
-      if (!isValidOutput(data.output)) throw ArgumentError('Output is invalid');
-      data.hash = data.output.sublist(3, 23);
+    } else if (data!.output != null) {
+      if (!isValidOutput(data!.output!)) throw ArgumentError('Output is invalid');
+      data!.hash = data!.output!.sublist(3, 23);
       _getDataFromHash();
-    } else if (data.pubkey != null) {
-      data.hash = hash160(data.pubkey);
+    } else if (data!.pubkey != null) {
+      data!.hash = hash160(data!.pubkey!);
       _getDataFromHash();
       _getDataFromChunk();
-    } else if (data.input != null) {
-      var _chunks = bscript.decompile(data.input);
+    } else if (data!.input != null) {
+      var _chunks = bscript.decompile(data!.input)!;
       _getDataFromChunk(_chunks);
       if (_chunks.length != 2) throw ArgumentError('Input is invalid');
       if (!bscript.isCanonicalScriptSignature(_chunks[0])) {
         throw ArgumentError('Input has invalid signature');
       }
       if (!isPoint(_chunks[1])) throw ArgumentError('Input has invalid pubkey');
-      data.witness = [];
+      data!.witness = [];
     } else {
       throw ArgumentError('Not enough data');
     }
   }
 
-  void _getDataFromChunk([List<dynamic> _chunks]) {
-    if (data.pubkey == null && _chunks != null) {
-      data.pubkey = (_chunks[1] is int) ? Uint8List.fromList([_chunks[1]]) : _chunks[1];
-      data.hash = hash160(data.pubkey);
+  void _getDataFromChunk([List<dynamic>? _chunks]) {
+    if (data!.pubkey == null && _chunks != null) {
+      data!.pubkey =
+          (_chunks[1] is int) ? Uint8List.fromList([_chunks[1]]) : _chunks[1];
+      data!.hash = hash160(data!.pubkey!);
       _getDataFromHash();
     }
-    if (data.signature == null && _chunks != null) {
-      data.signature = (_chunks[0] is int) ? Uint8List.fromList([_chunks[0]]) : _chunks[0];
+    if (data!.signature == null && _chunks != null) {
+      data!.signature =
+          (_chunks[0] is int) ? Uint8List.fromList([_chunks[0]]) : _chunks[0];
     }
-    if (data.input == null && data.pubkey != null && data.signature != null) {
-      data.input = bscript.compile([data.signature, data.pubkey]);
+    if (data!.input == null && data!.pubkey != null && data!.signature != null) {
+      data!.input = bscript.compile([data!.signature, data!.pubkey]);
     }
   }
 
   void _getDataFromHash() {
-    if (data.address == null) {
+    if (data!.address == null) {
       final payload = Uint8List(21);
       payload.buffer.asByteData().setUint8(0, network.pubKeyHash);
-      payload.setRange(1, payload.length, data.hash);
-      data.address = bs58check.encode(payload);
+      payload.setRange(1, payload.length, data!.hash!);
+      data!.address = bs58check.encode(payload);
     }
-    data.output ??= bscript.compile([OPS['OP_DUP'], OPS['OP_HASH160'], data.hash, OPS['OP_EQUALVERIFY'], OPS['OP_CHECKSIG']]);
+    data!.output ??= bscript.compile([
+      OPS['OP_DUP'],
+      OPS['OP_HASH160'],
+      data!.hash,
+      OPS['OP_EQUALVERIFY'],
+      OPS['OP_CHECKSIG']
+    ]);
   }
 
   void _getDataFromAddress(String address) {
@@ -76,18 +84,31 @@ class P2PKH {
     if (version != network.pubKeyHash) {
       throw ArgumentError('Invalid version or Network mismatch');
     }
-    data.hash = payload.sublist(1);
-    if (data.hash.length != 20) throw ArgumentError('Invalid address');
+    data!.hash = payload.sublist(1);
+    if (data!.hash!.length != 20) throw ArgumentError('Invalid address');
   }
 }
 
 bool isValidOutput(Uint8List data) {
-  return data.length == 25 && data[0] == OPS['OP_DUP'] && data[1] == OPS['OP_HASH160'] && data[2] == 0x14 && data[23] == OPS['OP_EQUALVERIFY'] && data[24] == OPS['OP_CHECKSIG'];
+  return data.length == 25 &&
+      data[0] == OPS['OP_DUP'] &&
+      data[1] == OPS['OP_HASH160'] &&
+      data[2] == 0x14 &&
+      data[23] == OPS['OP_EQUALVERIFY'] &&
+      data[24] == OPS['OP_CHECKSIG'];
 }
 
 // Backward compatibility
-@Deprecated("The 'P2PKHData' class is deprecated. Use the 'PaymentData' package instead.")
+@Deprecated(
+    "The 'P2PKHData' class is deprecated. Use the 'PaymentData' package instead.")
 class P2PKHData extends PaymentData {
   P2PKHData({address, hash, output, pubkey, input, signature, witness})
-      : super(address: address, hash: hash, output: output, pubkey: pubkey, input: input, signature: signature, witness: witness);
+      : super(
+            address: address,
+            hash: hash,
+            output: output,
+            pubkey: pubkey,
+            input: input,
+            signature: signature,
+            witness: witness);
 }
