@@ -2,51 +2,51 @@ import 'dart:typed_data';
 import 'constants/op.dart';
 
 class DecodedPushData {
-  int opcode;
-  int number;
-  int size;
+  int? opcode;
+  int? number;
+  int? size;
   DecodedPushData({this.opcode, this.number, this.size});
 }
 
 class EncodedPushData {
-  int size;
-  Uint8List buffer;
+  int? size;
+  Uint8List? buffer;
 
   EncodedPushData({this.size, this.buffer});
 }
 
-EncodedPushData encode(Uint8List buffer, number, offset) {
+EncodedPushData encode(Uint8List? buffer, number, offset) {
   var size = encodingLength(number);
   // ~6 bit
   if (size == 1) {
-    buffer.buffer.asByteData().setUint8(offset, number);
+    buffer!.buffer.asByteData().setUint8(offset, number);
 
     // 8 bit
   } else if (size == 2) {
-    buffer.buffer.asByteData().setUint8(offset, OPS['OP_PUSHDATA1']);
+    buffer!.buffer.asByteData().setUint8(offset, OPS['OP_PUSHDATA1']!);
     buffer.buffer.asByteData().setUint8(offset + 1, number);
 
     // 16 bit
   } else if (size == 3) {
-    buffer.buffer.asByteData().setUint8(offset, OPS['OP_PUSHDATA2']);
+    buffer!.buffer.asByteData().setUint8(offset, OPS['OP_PUSHDATA2']!);
     buffer.buffer.asByteData().setUint16(offset + 1, number, Endian.little);
 
     // 32 bit
   } else {
-    buffer.buffer.asByteData().setUint8(offset, OPS['OP_PUSHDATA4']);
+    buffer!.buffer.asByteData().setUint8(offset, OPS['OP_PUSHDATA4']!);
     buffer.buffer.asByteData().setUint32(offset + 1, number, Endian.little);
   }
 
   return EncodedPushData(size: size, buffer: buffer);
 }
 
-DecodedPushData decode(Uint8List bf, int offset) {
+DecodedPushData? decode(Uint8List bf, int offset) {
   var buffer = bf.buffer;
   var opcode = buffer.asByteData().getUint8(offset);
   int number, size;
 
   // ~6 bit
-  if (opcode < OPS['OP_PUSHDATA1']) {
+  if (opcode < OPS['OP_PUSHDATA1']!) {
     number = opcode;
     size = 1;
 
