@@ -43,7 +43,7 @@ TransactionBuilder constructSign(f, TransactionBuilder txb) {
   return txb;
 }
 
-TransactionBuilder construct(f, [bool dontSign]) {
+TransactionBuilder construct(f, [bool? dontSign]) {
   final network = networks[f['network']];
   final txb = TransactionBuilder(network: network);
   if (f['version'] != null) txb.setVersion(f['version']);
@@ -83,8 +83,8 @@ void main() {
   final fixtures = json.decode(File('test/fixtures/transaction_builder.json')
       .readAsStringSync(encoding: utf8));
   group('TransactionBuilder', () {
-    final keyPair = ECPair.fromPrivateKey(HEX.decode(
-        '0000000000000000000000000000000000000000000000000000000000000001'));
+    final keyPair = ECPair.fromPrivateKey(Uint8List.fromList(HEX.decode(
+        '0000000000000000000000000000000000000000000000000000000000000001')));
     final scripts = [
       '1BgGZ9tcN4rm9KBzDn7KprQz87SZ26SAMH',
       '1cMh228HTCiwS8ZsaakH8A8wze1JR5ZsP'
@@ -144,27 +144,27 @@ void main() {
         });
     });
     group('addInput', () {
-      TransactionBuilder txb;
+      late TransactionBuilder txb;
       setUp(() {
         txb = TransactionBuilder();
       });
       test('accepts a txHash, index [and sequence number]', () {
         final vin = txb.addInput(txHash, 1, 54);
         expect(vin, 0);
-        final txIn = txb.tx.ins[0];
+        final txIn = txb.tx!.ins[0];
         expect(txIn.hash, txHash);
         expect(txIn.index, 1);
         expect(txIn.sequence, 54);
-        expect(txb.inputs[0].prevOutScript, null);
+        expect(txb.inputs![0].prevOutScript, null);
       });
       test('accepts a txHash, index [, sequence number and scriptPubKey]', () {
         final vin = txb.addInput(txHash, 1, 54, scripts.elementAt(1));
         expect(vin, 0);
-        final txIn = txb.tx.ins[0];
+        final txIn = txb.tx!.ins[0];
         expect(txIn.hash, txHash);
         expect(txIn.index, 1);
         expect(txIn.sequence, 54);
-        expect(txb.inputs[0].prevOutScript, scripts.elementAt(1));
+        expect(txb.inputs![0].prevOutScript, scripts.elementAt(1));
       });
       test('accepts a prevTx, index [and sequence number]', () {
         final prevTx = Transaction();
@@ -174,11 +174,11 @@ void main() {
         final vin = txb.addInput(prevTx, 1, 54);
         expect(vin, 0);
 
-        final txIn = txb.tx.ins[0];
+        final txIn = txb.tx!.ins[0];
         expect(txIn.hash, prevTx.getHash());
         expect(txIn.index, 1);
         expect(txIn.sequence, 54);
-        expect(txb.inputs[0].prevOutScript, scripts.elementAt(1));
+        expect(txb.inputs![0].prevOutScript, scripts.elementAt(1));
       });
       test('returns the input index', () {
         expect(txb.addInput(txHash, 0), 0);
@@ -199,23 +199,23 @@ void main() {
       });
     });
     group('addOutput', () {
-      TransactionBuilder txb;
+      late TransactionBuilder txb;
       setUp(() {
         txb = TransactionBuilder();
       });
       test('accepts an address string and value', () {
         final address =
-            P2PKH(data: PaymentData(pubkey: keyPair.publicKey)).data.address;
+            P2PKH(data: PaymentData(pubkey: keyPair.publicKey)).data!.address;
         final vout = txb.addOutput(address, 1000);
         expect(vout, 0);
-        final txout = txb.tx.outs[0];
+        final txout = txb.tx!.outs[0];
         expect(txout.script, scripts.elementAt(0));
         expect(txout.value, 1000);
       });
       test('accepts a ScriptPubKey and value', () {
         final vout = txb.addOutput(scripts.elementAt(0), 1000);
         expect(vout, 0);
-        final txout = txb.tx.outs[0];
+        final txout = txb.tx!.outs[0];
         expect(txout.script, scripts.elementAt(0));
         expect(txout.value, 1000);
       });
