@@ -4,9 +4,9 @@ import '../../src/crypto.dart';
 import 'varuint.dart';
 import '../../src/models/networks.dart';
 
-Uint8List magicHash(String message, [NetworkType network]) {
+Uint8List magicHash(String message, [NetworkType? network]) {
   network = network ?? bitcoin;
-  Uint8List messagePrefix = utf8.encode(network.messagePrefix);
+  Uint8List messagePrefix = Uint8List.fromList(utf8.encode(network.messagePrefix));
   var messageVISize = encodingLength(message.length);
   var length = messagePrefix.length + messageVISize + message.length;
   var buffer = Uint8List(length);
